Rename `common_header.php` to `header.php`
